# server/docs/swagger.yaml
definitions:
...
  hotelSys.SysHotels:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      manager:
        type: string
      mark:
        type: string
      name:
        type: string
      phone:
        type: string
      platform:
        type: string
      settlementType:
        type: string
      signatory:
        type: string
      transferAccount:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      user_id:
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /sysHotels/createSysHotels:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建SysHotels
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/hotelSys.SysHotels'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建SysHotels
      tags:
      - SysHotels
  /sysHotels/deleteSysHotels:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除SysHotels
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/hotelSys.SysHotels'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除SysHotels
      tags:
      - SysHotels
  /sysHotels/deleteSysHotelsByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除SysHotels
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除SysHotels
      tags:
      - SysHotels
  /sysHotels/findSysHotels:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: manager
        type: string
      - in: query
        name: mark
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: phone
        type: string
      - in: query
        name: platform
        type: string
      - in: query
        name: settlementType
        type: string
      - in: query
        name: signatory
        type: string
      - in: query
        name: transferAccount
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询SysHotels
      tags:
      - SysHotels
  /sysHotels/getSysHotelsList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 主键ID
        in: query
        name: id
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: manager
        type: string
      - in: query
        name: mark
        type: string
      - in: query
        name: name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: phone
        type: string
      - in: query
        name: platform
        type: string
      - in: query
        name: settlementType
        type: string
      - in: query
        name: signatory
        type: string
      - in: query
        name: transferAccount
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取SysHotels列表
      tags:
      - SysHotels
  /sysHotels/updateSysHotels:
    post:
      consumes:
      - application/json
      parameters:
      - description: 更新SysHotels
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/hotelSys.SysHotels'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新SysHotels
      tags:
      - SysHotels
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"